# services/order-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Order API for the event-driven commerce practice project.\nIt persists orders, emits outbox events, and reflects asynchronous payment outcomes.",
        "title": "Go Event Commerce Order API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/livez": {
            "get": {
                "description": "Returns OK when the process is alive.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Liveness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.HealthResponseDoc"
                        }
                    }
                }
            }
        },
        "/orders": {
            "post": {
                "description": "Creates a new order, persists an outbox event, and publishes order.created.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "description": "Order payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateOrderRequestDoc"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.OrderResponseDoc"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/handler.OrderAcceptedResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponseDoc"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "description": "Returns the latest order state. Status can change asynchronously after payment processing.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.OrderResponseDoc"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponseDoc"
                        }
                    }
                }
            }
        },
        "/readyz": {
            "get": {
                "description": "Returns ready when dependencies are reachable.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Readiness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.HealthResponseDoc"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.ReadinessFailureResponseDoc"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.CreateOrderRequestDoc": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 129.99
                },
                "customer_id": {
                    "type": "string",
                    "example": "customer-001"
                }
            }
        },
        "handler.ErrorResponseDoc": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Key: 'CreateOrderRequestDoc.Amount' Error:Field validation for 'Amount' failed on the 'gt' tag"
                },
                "error": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
        "handler.HealthResponseDoc": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "handler.OrderAcceptedResponseDoc": {
            "type": "object",
            "properties": {
                "order": {
                    "$ref": "#/definitions/model.Order"
                },
                "warning": {
                    "type": "string",
                    "example": "order persisted but event publish failed"
                }
            }
        },
        "handler.OrderResponseDoc": {
            "type": "object",
            "properties": {
                "order": {
                    "$ref": "#/definitions/model.Order"
                }
            }
        },
        "handler.ReadinessFailureResponseDoc": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "dial tcp 127.0.0.1:5432: connect: connection refused"
                },
                "status": {
                    "type": "string",
                    "example": "not ready"
                }
            }
        },
        "model.Order": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}